Added option for fuzzy identifier matching. Closes #6.
It takes care of use cases where identifier in stack trace is application name, but addresses at the bottom use unique bundle identifier. For example "Startupizer from GentleBytes" vs. "com.gentlebytes.startupizer.from.gentlebytes". In this case it tries to match both, the bundle identifier and application name.

Note: for the moment this option need to be explicitly opted in via `--fuzzy` command line argument. I only got one report where this is needed so far so am not sure if it's particular use case, or general trend for future crash logs (this particular crash was on 10.13.1 which is about 1 month old at the moment of this writing). Because it may result in invalid matches, decided to keep it off by default for now.