Added automatic version and build number setup.
It uses build phase where it uses last git tag as version and number of commits as build number. Gathered info is copied as global constants to autogenerated file and used in `Options`.

Note this autogenerated file is ignored by git - it’ll be generated on the first run anyway.